Update docs and .gitignore
Our npm script to generate the docs outputs them into a "docs"
directory but it is ignored. When switching to the gh-pages branch to
update documentation, you can copy all files from the docs directory
into the main directory of the branch.  We want to ignore the "docs"
directory here to so we don't duplicate the pages.